% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGAM.R
\name{abr_breaks.hgam}
\alias{abr_breaks.hgam}
\title{Extracts breakpoints from an HGAM model}
\usage{
abr_breaks.hgam(
  abr_fitobj,
  breaktype = NULL,
  stepsize = NULL,
  delta = NULL,
  nsims = NULL,
  transform = NULL,
  aggregate = NULL,
  CI = FALSE
)
}
\arguments{
\item{abr_fitobj}{A previously fitted}
}
\value{
A list containing \code{seed}, the supplied seed, \code{initial_state}, the
initial state of the RNG before the seed was set, and \code{kind}, the type of
RNG used, as returned by \code{\link[base:Random]{base::RNGkind()}}.
}
\description{
Extracts breakpoints from an HGAM model
}
