% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_linear.R
\name{linear_trend}
\alias{linear_trend}
\title{Linear trend response model}
\usage{
linear_trend(t, start_value = 0, end_value = 1, ...)
}
\arguments{
\item{t}{numeric; vector of time points.}

\item{start_value, end_value}{numeric vectors of length 1; the start and end
values for the linear trend.}

\item{...}{other arguments. Ignored here.}
}
\description{
Linear trend response model
}
