% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{seed_rng}
\alias{seed_rng}
\title{Seed the RNG state possibly with a user supplied seed}
\usage{
seed_rng(seed = NULL)
}
\arguments{
\item{seed}{numeric seed for RNG. See \code{\link[base:set.seed]{base::set.seed()}}.}
}
\value{
A list containing \code{seed}, the supplied seed, \code{initial_state},
the initial state of the RNG before the seed was set, and
\code{kind}, the type of RNG used, as returned by \code{\link[base:RNGkind]{base::RNGkind()}}.
}
\description{
Seed the RNG state possibly with a user supplied seed
}
