% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_troph_triangle.R
\name{sim_troph_triangle}
\alias{sim_troph_triangle}
\title{Trophic triangle model}
\usage{
sim_troph_triangle(
  time_out = 100,
  n_steps = c(100),
  measurement_sigma = c(0, 0, 0),
  harvest_rates = create_driver_function(change_times = 50, parm_values = c(0, 0.25),
    interpolation = "constant"),
  pred_overlap = create_driver_function(change_times = 50, parm_values = c(1, 1),
    interpolation = "constant")
)
}
\arguments{
\item{time_out}{ToDo}

\item{n_steps}{ToDo}

\item{measurement_sigma}{ToDo}

\item{harvest_start}{ToDo}

\item{harvest_end}{ToDo}

\item{harvest_shape}{ToDo}

\item{pred_overlap_start}{ToDo}

\item{pred_overlap_end}{ToDo}

\item{pred_overlap_shape}{ToDo}
}
\description{
Trophic triangle model
}
\author{
Eric J. Pedersen
}
