% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_linear.R
\name{simulate_linear_trend}
\alias{simulate_linear_trend}
\title{Simulate data from a linear trend model}
\usage{
simulate_linear_trend(
  t,
  start_value = 0,
  end_value = 1,
  sampling_distribution = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{t}{numeric; vector of time points.}

\item{start_value, end_value}{numeric vectors of length 1; the start and end
values for the linear trend.}

\item{sampling_distribution}{function; a random number generating function,
which takes as it's first argument the number of observations to sample.
The second argument should be the expected value. The default, if nothing
is supplied, is \code{\link[stats:Normal]{stats::rnorm()}}.}

\item{seed}{numeric; a seed for the simulation.}

\item{...}{additional arguments that will be passed to
\code{sampling_distribution}.}
}
\description{
Simulate data from a linear trend model
}
