% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_step.R
\name{simulate_step_trend}
\alias{simulate_step_trend}
\title{Simulate data from a linear trend model}
\usage{
simulate_step_trend(
  t,
  change_points,
  means,
  sampling_distribution = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{t}{numeric; vector of time points.}

\item{change_points}{numeric; vector of change points, within \code{t}}

\item{means}{numeric; vector of means for the regimes implied by
\code{change_points}. Must be of length \code{length(change_points) + 1}.}

\item{sampling_distribution}{function; a random number generating function,
which takes as it's first argument the number of observations to sample.
The second argument should be the expected value. The default, if nothing
is supplied, is \code{\link[stats:rnorm]{stats::rnorm()}}.}

\item{seed}{numeric; a seed for the simulation.}

\item{...}{additional arguments that will be passed to
\code{sampling_distribution}.}
}
\description{
Simulate data from a linear trend model
}
\examples{
\dontshow{
set.seed(1)
op <- options(digits = 3, cli.unicode = FALSE)
}
sims <- simulate_step_trend(1:100, change_points = c(25, 75),
                            means = c(2, 8, 4))
sims

library("ggplot2")
ggplot(sims, aes(x = t, y = y)) +
  geom_point() +
  geom_step(aes(y = trend))
\dontshow{options(op)}
}
