% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_step.R
\name{step_trend}
\alias{step_trend}
\title{Step change trend response model}
\usage{
step_trend(t, change_points, means, ...)
}
\arguments{
\item{t}{numeric; vector of time points.}

\item{...}{other arguments. Ignored here.}

\item{start_value, end_value}{numeric vectors of length 1; the start and end
values for the linear trend.}
}
\description{
Step change trend response model
}
\examples{
\dontshow{
set.seed(1)
op <- options(digits = 3, cli.unicode = FALSE)
}
step_trend(1:100, change_points = c(25, 75), means = c(2, 8, 4))
\dontshow{options(op)}
}
