% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_troph_triangle.R
\name{sim_troph_triangle}
\alias{sim_troph_triangle}
\title{Trophic triangle model}
\usage{
sim_troph_triangle(
  time_out,
  n_steps,
  measurement_sigma = c(0, 0, 0),
  harvest_start = 0,
  harvest_end = 0.25,
  harvest_shape = 0,
  pred_overlap_start = 1,
  pred_overlap_end = 1,
  pred_overlap_shape = 0
)
}
\arguments{
\item{time_out}{}

\item{n_steps}{}

\item{measurement_sigma}{}

\item{harvest_start}{}

\item{harvest_end}{}

\item{harvest_shape}{}

\item{pred_overlap_start}{}

\item{pred_overlap_end}{}

\item{pred_overlap_shape}{}
}
\description{
Trophic triangle model
}
\author{
Eric J. Pedersen
}
