% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_step.R
\name{step_trend}
\alias{step_trend}
\title{Step change trend response model}
\usage{
step_trend(t, change_points, means, ...)
}
\arguments{
\item{t}{numeric; vector of time points.}

\item{change_points}{numeric; vector of change points, within \code{t}}

\item{means}{numeric; vector of means for the regimes implied by
\code{change_points}. Must be of length \code{length(change_points) + 1}.}

\item{...}{other arguments. Ignored here.}
}
\description{
Step change trend response model
}
\examples{
\dontshow{
set.seed(1)
op <- options(digits = 3, cli.unicode = FALSE)
}
sims <- step_trend(1:100, change_points = c(25, 75), means = c(2, 8, 4))
sims

library("ggplot2")
ggplot(sims, aes(x = time, y = trend)) +
  geom_step()
\dontshow{options(op)}
}
